backend/pyproject.toml
[project]
name = "teachable-agent"
version = "0.1.0"
description = "An agent that can improve its execution capability through user teaching."
authors = [
    { name = "Shawn Zhang", email = "shawnzhang31@gmail.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.25",
    "langchain-deepseek>=0.1.3",
    "langchain-openai>=0.3.22",
    "langgraph>=0.4.8",
    "langgraph-cli[inmem]>=0.3.1",
    "langmem>=0.0.27",
    "python-dotenv>=1.1.0",
]
